+++ CHANGELOG.md
@@ -39,6 +39,12 @@
     * Change `PayPalClient#tokenizePayPalAccount` parameters
     * Add `PayPalClient#handleReturnToAppFromBrowser`
 
+## unreleased 
+
+* GooglePay
+  * Bump `play-services-wallet` version to `19.2.1`
+  * Add `totalPriceLabel` to `GooglePayRequest`
+
 ## 4.39.0 (2023-10-16)
 
 * BraintreeCore